offload 新增音量dfx 告警修复

+++ services/audio_service/common/src/volume_tools.cpp
@@ -423,8 +423,8 @@ static void CountS32Volume(const BufferDesc &buffer, AudioChannel channel, Chann
         return;
     }
     for (size_t index = 0; index < channel; index++) {
-        volSums[index] /= static_cast<int32_t>(size);
-        volMaps.volStart[index] = volSums[index];
+        volSums[index] /= static_cast<int64_t>(size);
+        volMaps.volStart[index] =  static_cast<int32_t>(volSums[index]);
     }
     return;
 }
